Fix file locations so that both tests and CLI work
