+++ loconotion/tests/test_parser.py
@@ -1,5 +1,5 @@
 import pytest
-from loconotion.notionparser import Parser
+from modules.notionparser import Parser
 
 def test_parse_sample_page():
     config={"page": "https://www.notion.so/Loconotion-Example-Page-03c403f4fdc94cc1b315b9469a8950ef"}
